Move JS code into 'modules' so it's easier to track dependencies'
