+++ static/js/browser/kg_template.jsx
