+++ static/js/browser/kg_template.test.jsx
