+++ static/js/browser/plot.js
