+++ static/js/browser/view.js
@@ -16,7 +16,7 @@
 
 const _ = require("lodash");
 const Cookies = require("js-cookie");
-const util = require("./util.js");
+const util = require("../shared/util.js");
 
 const TEXT = "text";
