+++ static/js/place/place.ts
