+++ static/js/shared/data_fetcher.test.ts
@@ -17,7 +17,7 @@
 import axios from "axios";
 
 import { fetchStatsData } from "./data_fetcher";
-import { DataGroup } from "./chart/base";
+import { DataGroup } from "../chart/base";
 
 jest.mock("axios");
 const mockedAxios = axios as jest.Mocked<typeof axios>;
