+++ static/js/shared/stats_var.ts
