+++ static/js/shared/util.js
