+++ static/js/tools/globals.js
